Update code to print carriage ret instead of newline
In old computers, newline will drop you to the next line, but instead of
beginning at the leftmost position, you begin at the position that is
just below where you were. The carriage return is sort of like how it
was with typewriters: it newlines and ALSO moves all the way to the
left.

This commit will cause us to use the \r rather than the \n character for
this reason.